Zipped Big Data Project "Big Grega"
If you are accessing our code VIA Github, load all the library on line 14 and create all the file paths and folders starting from line 19. Skip to line 209 in our code our files were to large to upload onto Github, so we only uploaded the cleaned file with the data we actually used.