Zipped Big Data Project "Big Grega"
If you are accessing our code VIA Github, load all the library on line 16 and create all the file paths and folders starting from line 21. Skip to line 210 in our code our files were too large to upload onto Github, so we only uploaded the cleaned file with the data we actually used.